% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{coefficient_to_variable}
\alias{coefficient_to_variable}
\title{Converts apollo coefficient_name (b_) to variable name.}
\usage{
coefficient_to_variable(coef_name)
}
\arguments{
\item{coef_name}{}
}
\value{
variable name
}
\description{
Converts apollo coefficient_name (b_) to variable name.
}
\examples{
variable <- coefficient_to_variable("b_age")
}
