% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{compare_models}
\alias{compare_models}
\title{Compare feature importance of models using the Wald test}
\usage{
compare_models()
}
\description{
Uses the Wald test to pairwise compare features in both models. The Wald test is defined as
\eqn{W = \frac {(\beta_{1} - \beta_{2}) ^2}{\operatorname {se}(\beta_{1})^2 + \operatorname {se}(\beta_{2})^2} }
}
