% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_criteria_table}
\alias{create_criteria_table}
\title{Create Criteria table}
\usage{
create_criteria_table()
}
\value{
gt table of the criteria (Table 1)
}
\description{
Displays the criteria and associated levels used in the discrete choice experiment. Saves
the pivot table as \code{extdata/tables/table_criteria.html}.
}
\examples{
baitlist::create_criteria_table()
}
