% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_group_comparison_tables}
\alias{create_group_comparison_tables}
\title{Create group comparison table}
\usage{
create_group_comparison_tables()
}
\value{
list of gt tables
}
\description{
Creates tables for both the binary and multinomial models comparing
coefficients weights with both groups using the Wald test.
Saves the table as \code{data/tables/table_group_comparison.html}.
}
\examples{
tables <- baitlist::create_group_comparison_tables()
tables$binary
}
