% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{criteria_table}
\alias{criteria_table}
\title{Create Criteria table}
\usage{
criteria_table()
}
\value{
Tibble containing criteria
}
\description{
Displays the criteria and associated levels used in the discrete choice experiment. Saves
the pivot table as \code{data/tables/criteria.csv}.
}
\examples{
criteria_table()
}
