% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance_all}
\alias{feature_importance_all}
\title{Calculates feature importance for all BAIT model types and groups using
standardized coefficients and maximum utility contribution.}
\usage{
feature_importance_all()
}
\description{
Standardized Coefficients (\eqn{\beta ^{\ast }={\frac {s_{x}}{s_{y}}}\beta}) as
percentage of total.
}
\details{
Maximum Utility Contribution is calculated by multiplying the criteria level range
(difference between the highest and lowest level) with the absolute value of the
criterion weight. To calculate a percentage, the maximum utility contributions
are summed.

Feature importance is saved for each model as csv files in the \code{inst/extdata} folder.
}
\examples{
baitlist::feature_importance_all()
}
