% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance_binary}
\alias{feature_importance_binary}
\title{Calculate feature importance}
\usage{
feature_importance_binary()
}
\description{
Calculates feature importance of the binary BAIT models using standardized coefficients
(\eqn{\beta ^{\ast }={\frac {s_{x}}{s_{y}}}\beta}) as percentage of total.
Feature importance is saved for each model as csv files in the \code{data} folder.
}
\examples{
feature_importance_binary()
}
