% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled.R
\name{fit_pooled_interaction_multinomial}
\alias{fit_pooled_interaction_multinomial}
\title{Fits a pooled multinomial model with interaction effects using data
from both \code{group1} and \code{group2} to allow comparing difference in
coefficients between groups}
\usage{
fit_pooled_interaction_multinomial(group1, group2)
}
\arguments{
\item{group1}{First Group}

\item{group2}{Second group}
}
\value{
Apollo model object
}
\description{
Fits a pooled multinomial model with interaction effects using data
from both \code{group1} and \code{group2} to allow comparing difference in
coefficients between groups
}
\examples{
fit_pooled_interaction_multinomial('aumc', 'olvg')
}
