% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled.R
\name{fit_pooled_models}
\alias{fit_pooled_models}
\title{Fit pooled models with interaction effects for group-wise comparisons to allow
comparing coefficients for Amsterdam UMC vs. OLVG and
Intensivists vs. Fellows for both binary and multinomial models}
\usage{
fit_pooled_models()
}
\description{
Fit pooled models with interaction effects for group-wise comparisons to allow
comparing coefficients for Amsterdam UMC vs. OLVG and
Intensivists vs. Fellows for both binary and multinomial models
}
\examples{
baitlist::fit_pooled_models()
}
