% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_responses.R
\name{load_responses}
\alias{load_responses}
\title{Load responses
Reads data from the data/responses.xlsx file for specific groups or all if not specified.}
\usage{
load_responses(group = "aggregate")
}
\arguments{
\item{group}{The group to filter the data on.
Valid values: 'aumc', 'olvg', 'intensivists', 'fellows'. If omitted all responses are returned.}
}
\value{
data.frame (tibble) containing responses for the specified group
}
\description{
Load responses
Reads data from the data/responses.xlsx file for specific groups or all if not specified.
}
\examples{
data_fellows <- load_responses('fellows')

}
