% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_pooled_group_comparison}
\alias{plot_pooled_group_comparison}
\title{Plots group comparison of coefficient weights}
\usage{
plot_pooled_group_comparison()
}
\value{
list of ggplots
}
\description{
Creates plots with coefficients weights and significance levels comparing both groups using the Wald test.
Saves the plots as png and svg in the \verb{data\\figures} folder.
}
\examples{
baitlist::plot_pooled_group_comparison()
}
