% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_relative_importance}
\alias{plot_relative_importance}
\title{Plot Relative Importance}
\usage{
plot_relative_importance()
}
\value{
list of ggplots
}
\description{
Creates plots showing the relative factor importance using Maximum Utility Contribution for the aggregate
model for both the binary and multinomial models of all participants and subgroups.
Saves the plot as png and svg in the \verb{extdata\\figures} folder.
}
\examples{
baitlist::plot_relative_importance()
}
