% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict choice based on patient characteristics and family values}
\usage{
predict(
  expected_los = 0,
  clinical_situation = 0,
  age = 0,
  frailty = 0,
  life_expectancy = 0,
  suffering = 0,
  disability_cardiovascular = 0,
  disability_pulmonary = 0,
  disability_renal = 0,
  disability_neurological = 0,
  disability_gastrointestinal = 0,
  family_values = 0,
  type = "binary",
  group = "aggregate"
)
}
\arguments{
\item{expected_los}{Expected length of stay}

\item{clinical_situation}{Clinical Situation}

\item{age}{Age}

\item{frailty}{Frailty}

\item{life_expectancy}{Life Expectancy}

\item{suffering}{Suffering}

\item{disability_cardiovascular}{Expected cardiovascular disability after discharge}

\item{disability_pulmonary}{Expected pulmonary disability after discharge}

\item{disability_renal}{Expected renal disability after discharge}

\item{disability_neurological}{Expected neurological disability after discharge}

\item{disability_gastrointestinal}{Expected gastrointestinal disability after discharge}

\item{family_values}{Family values}

\item{type}{Either 'binary' or 'multinomial'. Default: 'binary'.}

\item{group}{Either 'aggregate', 'aumc', 'olvg', 'intensivists' or 'fellows'. Default: 'aggregate'}
}
\value{
Returns tibble containing prediction
}
\description{
Used the \code{apollo} package to predict the choice based on the fitted models
}
\examples{
predict(family_values = 1)
}
