% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{run_linearity_tests}
\alias{run_linearity_tests}
\title{Runs linearity tests for multi-level ordered categories.}
\usage{
run_linearity_tests()
}
\value{
list containing gt table ('summary_table') and combined plot ('dashboard')
}
\description{
Runs linearity tests for multi-level ordered categories.
}
\examples{
lin <- run_linearity_tests()
lin$summary_table
}
