% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald_interaction}
\alias{wald_interaction}
\title{Creates a dataframe containing the Joint Wald test and Wald tests for
individual coefficients in a pooled interaction model of two groups}
\usage{
wald_interaction(group1, group2, modeltype)
}
\arguments{
\item{group1}{First group}

\item{group2}{Second group}

\item{modeltype}{Model type: 'binary' or 'multinomial'}
}
\value{
dataframe
}
\description{
Creates a dataframe containing the Joint Wald test and Wald tests for
individual coefficients in a pooled interaction model of two groups
}
\examples{
baitlist::wald_interaction('intensivists', 'fellows', 'binary')
}
