% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_criteria_table}
\alias{create_criteria_table}
\title{Create Criteria table}
\usage{
create_criteria_table()
}
\value{
Tibble containing criteria
}
\description{
Displays the criteria and associated levels used in the discrete choice experiment. Saves
the pivot table as \code{data/tables/table_criteria.csv}.
}
\examples{
create_criteria_table()
}
