% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_group_comparison_table}
\alias{create_group_comparison_table}
\title{Create group comparison table}
\usage{
create_group_comparison_table()
}
\value{
gt table
}
\description{
Creates table comparing coefficients weights with both groups using the Wald test.
Saves the table as \code{data/tables/table_group_comparison.html}.
}
\examples{
create_group_comparison_table()
}
