% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_model_weights_tables}
\alias{create_model_weights_tables}
\title{Create Model Weights tables}
\usage{
create_model_weights_tables()
}
\value{
list of gt tables
}
\description{
Displays the model weights, constant, Null log-likelihood and estimated model log-likelihood and
adjusted rho-squared of the binary and multinomial models for each model separately and a large
combined table for across all groups of each model type.
Saves the tables as \verb{extdata/tables/<binary|multinomial>/model_weights_<group>.html} using the
gt package.
}
\examples{
tables <- baitlist::create_model_weights_tables()
tables$binary$aggregate
}
