% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_model_weights_tables}
\alias{create_model_weights_tables}
\title{Create Model Weights tables}
\usage{
create_model_weights_tables()
}
\value{
gt table of last model
}
\description{
Displays the model weights, constant, Null log-likelihood and estimated model log-likelihood and
adjusted rho-squared of the different models. Saves the table as \code{data/tables/model_weights_<.csv}.
}
\examples{
create_model_weights_tables()
}
