% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bait_binary.R
\name{fit_bait_binary}
\alias{fit_bait_binary}
\title{Fits the responses in a BAIT binary logistic regression model for each
respondent group}
\usage{
fit_bait_binary()
}
\description{
Uses the \code{apollo} package for modelling.
}
\examples{
fit_bait_binary()
}
