% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bait_multinomial.R
\name{fit_bait_multinomial}
\alias{fit_bait_multinomial}
\title{Fits the responses in a BAIT multinomial logistic regression model for each
respondent group}
\usage{
fit_bait_multinomial()
}
\description{
Uses the \code{apollo} package for modelling.
}
\examples{
fit_bait_multinomial()
}
