% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_relative_importance}
\alias{plot_relative_importance}
\title{Plot Relative Importance}
\usage{
plot_relative_importance()
}
\value{
ggplot object
}
\description{
Creates a plot showing the relative factor importance using Maximum Utility Contribution for the aggregate
model. Saves the plot as png and svg in the \verb{data\\figures} folder.
}
\examples{
plot_relative_importance()
}
