% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{standardized_coefficients}
\alias{standardized_coefficients}
\title{Calculate feature importance using standardized coefficients}
\usage{
standardized_coefficients(modeltype = "binary", group = "aggregate")
}
\arguments{
\item{modeltype}{Model type. Either 'binary' or 'multinomial'}

\item{group}{Participant group. One of the following 'aggregate', 'aumc', 'olvg', 'intensivists', 'fellows'}
}
\description{
Calculates feature importance of the BAIT models using standardized coefficients
(\eqn{\beta ^{\ast }={\frac {s_{x}}{s_{y}}}\beta}) as percentage of total.
Feature importance is saved for each model as csv files in the \code{data} folder.
}
\examples{
baitlist::standardized_coefficients()
}
